Updated example
