+++ examples/autoupdatergui/impleUpdaterGui/SimpleUpdaterGui.pro
