+++ examples/autoupdatergui/impleUpdaterGui/mainwindow.h
